refactor(pwa): extract DueDateCell, TaskRow, NewTaskRow from tasks.tsx
Finishes the tasks.tsx decomposition started earlier. Move the three large leaf
components into components/tasks/ (verbatim, behaviour-preserving): DueDateCell
(date/recurrence/reminders popover), TaskRow (~480-line row), and NewTaskRow
(inline add-row, exporting NewTaskInput for the container). tasks.tsx drops from
2509 → 1275 lines and is now just the Tasks container + data layer. Typecheck,
ESLint, and the Vitest suite all pass.
